Refactor initialize method
There's no need to put defaults in the parameters, it just
makes the method harder to read. In particular, the @silent
and @stepwise instance variables don't need defaults, since
if their corresponding ENV vars aren't used, they will simply
return `nil`, which is `false` in a boolean context, and `false`
is the desired default value.